+++ lib/karafka/web/pro/ui/views/explorer/explorer/_message.erb
