+++ lib/karafka/web/pro/ui/views/explorer/explorer/_no_topics.erb
