+++ lib/karafka/web/pro/ui/views/explorer/explorer/_partition_option.erb
