+++ lib/karafka/web/pro/ui/views/explorer/explorer/_selector.erb
@@ -7,7 +7,7 @@
     <%==
       each_partial(
         [nil] + @partitions_count.times.to_a,
-        'explorer/partition_option',
+        'explorer/explorer/partition_option',
         local: :partition
       )
     %>
