+++ lib/karafka/web/pro/ui/views/explorer/explorer/_topic.erb
