+++ lib/karafka/web/pro/ui/views/explorer/explorer/message/_too_big_to_be_displayed.erb
