+++ lib/karafka/web/pro/ui/views/explorer/explorer/messages/_headers.erb
