+++ lib/karafka/web/pro/ui/views/explorer/explorer/messages/_key.erb
