+++ lib/karafka/web/pro/ui/views/explorer/explorer/partition/_cleaned.erb
