+++ lib/karafka/web/pro/ui/views/explorer/explorer/partition/_empty.erb
