+++ lib/karafka/web/pro/ui/views/explorer/explorer/partition/_messages.erb
@@ -17,7 +17,7 @@
       <%==
         each_partial(
           @messages,
-          'explorer/message'
+          'explorer/explorer/message'
         )
       %>
     </tbody>
