+++ lib/karafka/web/pro/ui/views/explorer/explorer/topic.erb
@@ -3,17 +3,17 @@
 
 <% view_title @topic_id %>
 
-<%== partial 'explorer/topic/actions' %>
+<%== partial 'explorer/explorer/topic/actions' %>
 
 <div class="col-span-12">
   <% if @limited %>
     <div class="mb-6">
-      <%== partial('explorer/topic/limited') %>
+      <%== partial('explorer/explorer/topic/limited') %>
     </div>
   <% end %>
 
   <% if @messages.empty? && params.current_page == 1 %>
-    <%== partial 'explorer/topic/empty' %>
+    <%== partial 'explorer/explorer/topic/empty' %>
   <% elsif @messages.empty? %>
     <%== partial 'shared/no_paginated_data' %>
   <% else %>
@@ -29,6 +29,6 @@
       </p>
     <% end %>
 
-    <%== partial('explorer/partition/messages') %>
+    <%== partial('explorer/explorer/partition/messages') %>
   <% end %>
 </div>
