+++ lib/karafka/web/pro/ui/views/explorer/explorer/topic/_actions.erb
@@ -25,6 +25,6 @@
 
 <div class="col-span-8 mb-3">
   <section class="actions">
-    <%== partial 'explorer/selector' %>
+    <%== partial 'explorer/explorer/selector' %>
   </section>
 </div>
