+++ lib/karafka/web/pro/ui/views/explorer/explorer/topic/_empty.erb
