+++ lib/karafka/web/pro/ui/views/explorer/search/_breadcrumbs.erb
@@ -1,4 +1,4 @@
 <%# This code is part of Karafka Pro, a commercial component not licensed under LGPL. %>
 <%# See LICENSE for details. %>
 
-<%== partial 'explorer/breadcrumbs' %>
+<%== partial 'explorer/explorer/breadcrumbs' %>
