+++ lib/karafka/web/pro/ui/views/explorer/search/_fix_errors.erb
