+++ lib/karafka/web/pro/ui/views/explorer/search/_metadata.erb
