+++ lib/karafka/web/pro/ui/views/explorer/search/_no_results.erb
