+++ lib/karafka/web/pro/ui/views/explorer/search/_no_search_criteria.erb
