+++ lib/karafka/web/pro/ui/views/explorer/search/_search_modal.erb
