+++ lib/karafka/web/pro/ui/views/explorer/search/_timeout.erb
