+++ lib/karafka/web/pro/ui/views/scheduled_messages/explorer/topic.erb
@@ -5,17 +5,17 @@
 
 <% @selector_path_method = :scheduled_messages_explorer_path %>
 
-<%== partial 'explorer/topic/actions' %>
+<%== partial 'explorer/explorer/topic/actions' %>
 
 <div class="col-span-12">
   <% if @limited %>
     <div class="mb-6">
-      <%== partial('explorer/topic/limited') %>
+      <%== partial('explorer/explorer/topic/limited') %>
     </div>
   <% end %>
 
   <% if @messages.empty? && params.current_page == 1 %>
-    <%== partial 'explorer/topic/empty' %>
+    <%== partial 'explorer/explorer/topic/empty' %>
   <% elsif @messages.empty? %>
     <%== partial 'shared/no_paginated_data' %>
   <% else %>
