+++ lib/karafka/web/pro/ui/views/topics/index.erb
@@ -4,7 +4,7 @@
 <% view_title 'Topics' %>
 
 <% if @topics.empty? %>
-  <%== partial 'explorer/no_topics' %>
+  <%== partial 'explorer/explorer/no_topics' %>
 <% else %>
   <div class="topic-tiles">
     <%==
