+++ spec/lib/karafka/web/pro/ui/controllers/explorer/explorer_controller_spec.rb
