+++ spec/lib/karafka/web/pro/ui/controllers/explorer/search_controller_spec.rb
